basico

+++ src/pages/CreateSquadScreen.js
@@ -34,26 +34,32 @@ const CreateSquadScreen = () => {
         const min = 3;
         const max = 6;
 
-        if (countAll >= min && countAll <= max) {
-            await database()
-                .ref(`/grupo/${region}/${Date.now()}`)
-                .set({
-                    name: name,
-                    type: type,
-                    description: description,
-                    pokemons: pokemons,
-                    user: userToken,
-                    time: Date(),
-                })
-                .then(() => {
-                    dispatch(deleteAllPokemons());
-                    console.log('Data set.');
-                });
-            Alert.alert('Group created correctly')
-            navigation.navigate('GroupsScreen', { region: region })
+        if (name) {
+            if (countAll >= min && countAll <= max) {
+                await database()
+                    .ref(`/grupo/${region}/${Date.now()}`)
+                    .set({
+                        name: name,
+                        type: type,
+                        description: description,
+                        pokemons: pokemons,
+                        user: userToken,
+                        time: Date(),
+                    })
+                    .then(() => {
+                        dispatch(deleteAllPokemons());
+                        console.log('Data set.');
+                    });
+                Alert.alert('Group created correctly')
+                navigation.navigate('GroupsScreen', { region: region })
+            } else {
+                Alert.alert('Debes considerar Minimo 3 y Maximo 6 pokemones')
+            }
         } else {
-            Alert.alert('Debes considerar Minimo 3 y Maximo 6 pokemones')
+            Alert.alert('El campo nombre es Obligatorio')
         }
+
+
 
 
     }
